Dc 404 fimsctl dropdown choice (#451)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-404)

## Description
Create a fimsctl for actions. 
Needs to have an enabled field and a list of possible actions derived
from the internal actions vector.

So what this does is add some more JSON to the ESS instance pubs that
indicates if the maint_mode "tray" should allow a user to use an action.
### How to test. 
Do a get on ess_1.
`fims_send -m get -u /assets/ess/ess_1 -r /$$ | jq '.maint_actions_ctl'`
<--- This is the JSON this pr creates. In this JSON you will find the
name of each possible action that instance can undergo, and a return
value. Both of these are the same, but Annie mentions perhaps we end up
changing the name to be a human legible version of the return value.
Anywho, That is a different ticket. There is also the enabled true/false
which tells the UI if it should allow the user to use the ctl.

### Scope creep
Needed a way to clear an action when it is in a static state (not in
progress) so that the UI will stop rendering it.
This also sets up a clear action. To test this start an action and then
stop it during execution. Then do a get on the action. Then clear it and
it should be "reset" and "inactive".

So ...
`fims_send -m set -u /assets/ess/ess_1/actions/Calibration1/start true`
Wait for like 10 seconds.
`fims_send -m set -u /assets/ess/ess_1/actions/Calibration1/stop true`
`fims_send -m get -u /assets/ess/ess_1/actions/Calibration1 -r /$$` <---
note the indexes
`fims_send -m set -u /assets/ess/ess_1/actions/Calibration1/clear true`
<--- resets the indexes and marks as inactive
`fims_send -m get -u /assets/ess/ess_1/actions/Calibration1 -r /$$` <---
note the indexes

If you have the ui up you should see the stepper vanish on clear.

## How Has This Been Tested?
Manual testing on the new UI.