DC-421 | UI | Login Page Uplift UI to Design System V2 (#486)
## [DC-421 | UI | Login Page Uplift UI to Design System
V2](https://flexgen.atlassian.net/browse/DC-421)

## Description
Updates the graphic and color used on the login page of HOS Control and
Coordinate to match updated designs in Figma. Purely stylistic change,
should not affect any functionality. Also adds configs to the sandbox so
that the customer name and server name appear on the login page.

## How Has This Been Tested?
Manually tested in the UI