DC-440 | UI | Dark Mode Styling Errors (#482)
## [DC-440 | UI | Dark Mode Styling
Errors](https://flexgen.atlassian.net/browse/DC-440)

## Description
Updates the styling on the Statuses container on the configurable asset
pages to pull the correct color token in dark mode. Also updates the
version of storybook used in the application to incorporate a fix to the
DualOptionButton that appears on the dashboard, allowing users to switch
between card and table mode. This component should now appear correctly
in dark mode.

## How Has This Been Tested?
Manually tested in the UI