ACO-218-go_metrics-stops-publishing-outputs-and-responding-to-gets (#503)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-218)

## Description
go_metrics stops publishing outputs and responding to gets after running
for about 15-30 minutes on engie/libra when running v10.3.14. This PR
addresses that problem by reducing the use of global
map[string]interfaces{} when publishing and replying to messages.

## How Has This Been Tested?
Throwing lots and lots and lots of get requests at the system (of all
types of points, making sure to access every branch of the "gets" part
of the code).