CI-25 add device id to sim support systems (#488)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/CI-25)

## Description
Gcom_modbus_client now supports the use of multiple device_ids in a
single configuration document. As a result, we need to update the
support systems (echo config generator, config composer, and Map
Manager) to allow for this change. Configurations can be made in the old
format using the previous methods of config generation, but a new
`gcom_modbus` flag has been added to allow for configurations with
multiple device_ids.

## How Has This Been Tested?
Pulled several modbus_client configs from integration_dev and used them
to generate the echo and server configs. Ran diffs to make sure that the
old files are exactly the same as they were before.