Stop the Spam... Oops (#507)
It passed all the pytests though...

## [Link to JIRA Ticket](replacemewithlinktojiraticket)

## Description
In a recent merge I accidentally created a log that will spam while an
action is being run. The log was a quick last minute addition that
compiled and passed all the pytest. I'll try to be more watchful in the
future. My b.
## How Has This Been Tested?
Manually assure the log no longer spams.