set modbus connection timeout (#526)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-234)

## Description
Though modbus_client contains a "connection_timeout" field, this value
was used to set the fims timeout. We need it to set the modbus
connection timeout instead.

## How Has This Been Tested?
Set modbus_server delay (built using -DSERVER_DELAY) to 1500 ms and ran
the agt pv_1.json config against the server with a poll time of 6000 ms
(to account for the fact that all three sets of registers needed to come
back). This got rid of the connection timeout errors that we were seeing
without the "connection_timeout" flag working.