ACO-238-go_metrics-filters-with-string-comparisons (#539)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-238)

## Description
Filters containing strings did not remove the extra quotes from the
string. Thus, we were comparing `Online` to `"Online"` (for example).
The extra quotes would make these expressions not equal, with no way to
get rid of the extra quotes.

## How Has This Been Tested?
Added additional test case for filters with strings. (Represents the
particular use case that we have for it right now.)

Release notes: bugfix for go_metrics filters with strings not removing
extra quotations