DC-436 | UI | Sortable Coordinate Dashboard (#510)
## [DC-436](https://flexgen.atlassian.net/browse/DC-436)

## Description
Adds sorting behavior to the data tables in the Dashboard table view of
Coordinate and Control. The tables are sortable by any column header by
clicking on the arrow that appears next to the column title. Sorting
behavior switches between ascending/descending order with subsequent
clicks.

## How Has This Been Tested?
Manually tested in the UI

[DC-436]:
https://flexgen.atlassian.net/browse/DC-436?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ