DataMaps Version 1.0 into dev (#531)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-216)

## Description
This is version 1.0 of the dataMaps code to be released on the
ess_controller. Datamaps is a general term for the processes that the
ess controller uses to run arbitrary C/C++ code. The main use case for
this first iteration of datamaps is to run simulink auto-generated code
for EKF deployment on the ess_controller. Datamaps uses a combination of
the ess scheduler and threads to run code in such a way that the datamap
code does not interfere with other processes running on the ess
controller and does not significantly impact the overall performance of
the system.

In addition to the implementation of DataMaps, a small change was made
to the ess_controller source code in FlexPack.cpp line 2189 and in
VarMapUtils.cpp lines 3779, 3781, 3861, and 3863. These changes were
made because functions run on the ess controller were getting their
asset managers changed when they should not be getting changed.

Running the dataMaps code using Valgrind provides no memory leaks from
the dataMaps code itself. Certain SL files do have memory leaks, but all
of the files under the /ess_controller/SL_files/ folder are preliminary
and are only used for testing purposes.

Documentation for DataMaps will be on Confluence soon.

Note: This PR replaces the ACO-66 PR. Ignore that one and use this one
instead.

## How Has This Been Tested?
This has been rigorously tested under many different conditions with
many different configurations. Multiple thread have been tested,
multiple threads with multiple assetVars each, and multiple functions on
each assetVar. The first use case for dataMaps (EKF deployment) has not
yet been developed and has not been tested with that code. More testing
of the dataMaps infrastructure will come as more code uses it to run on
the ess controller.