Aco 254 go metrics direct set sparse option doesn t work (#581)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-254)

## Description
The "sparse" flag should allow direct sets to only be sent out when they
change. However, a missing ! was causing this flag not to work. Added
the ! where it should be and now it works again. (The other changes are
just from gopls adjusting the indentation and spacing.)

## How Has This Been Tested?
Manually, using `go_metrics/configs/config_doc_1.json` and sending the
same value to the system repeatedly.

Possible release notes: bugfix for direct_sets not working with "sparse"
flag