DC-409 Fix unescaped string from modbus (#563)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-409)

## Description
This patch implements some extra checks on fims messages during parsing
in the event a fims body is un-escaped. This ticket also fixes a current
known location of sending un-escaped strings in modbus client when a
FIMS get is sent to a modbus client uri, beginning with `/components`,
that does not exist.
The first origin of this error was from COPS, when listing modbus
clients in COPS (for the status page connection reporting) that did not
have heartbeats, COPS would attempt to send a FIMS get at the uri in
which a heartbeat did not exist. Modbus client would reply back with the
un-escaped FIMS body.

## How Has This Been Tested?
- COPS currently implements the fix to not send FIMS gets for a
heartbeat on every listed process. To have COPS send for a heartbeat
value, `requiredForHealthyStatus` must be set to true.
- In cops configuration, set this value to true for a given modbus
client listed process to force the error.
- with the fix in modbus_client.cpp, running a fims get at the uri
`/cops/heartbeat/modbus_client_example.json` should reply back with the
properly escaped string: "Modbus Client -> URI does not exist"