Aco 253 data map interface map fix (#611)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-253)

## Description
The concept behind running functions using datamaps is that you can run
your function on multiple different assetVars and have multiple
different instances of the same function on those multiple assetVars.
Previously, when running functions using datamaps, instances of
functions across assetVars were not properly differentiated and
incorrect configuration of an assetVar could lead to unexpected
behavior.
A solution has been implemented to ensure that functions on different
assetVars are running different models and data and inputs from one
model does not impact the other model. This is done by expanding the
dataMaps interface map to be 2 dimensional, using the assetVar name and
instance number when accessing model objects.

## How Has This Been Tested?
This has been tested using the same tests that were used before merging
the original DataMaps version 1.0 into dev and this branch has all the
same functionality.
When testing this on your machine, be sure to uncomment the datamap
files from the make file and uncomment references to those files in
dataMapCore.cpp.