DCM-208 dts does not shutdown if a database goes down (#600)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DCM-208)

## Description

Allows DTS to continue running and processing archives even if one of
the destination databases becomes unavailable. DTS will still shut down
if both databases become unavailable.

As a part of implementation, this PR also moves influxdb_client and
mongodb_client into the hybridos monorepo.

## How Has This Been Tested?

Created a docker test setup with a DTS container, and Influx container,
and a Mongo container. Manually tested within this setup:
Tested that DTS continues processing when Influx is stopped, and can
reconnect when Influx is restarted.
Tested that DTS continues processing when Mongo is stopped, and can
reconnect when Mongo is restarted.
Tested that DTS shuts down if both Influx and Mongo are stopped.
Tested that DTS can process archives when Influx is stopped at startup.
Tested that DTS can process archives when Mongo is stopped at startup.