ACO 239 go metrics conditional aggregation (#607)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-239)

## Description
This PR creates a new filter called "IndexFilter" which allows the user
to filter by index. Essentially, it does an index-wise comparison of the
filter variable to comparand, and extracts the matching elements. e.g.
"for each PCS, if that PCS's status, which pubbed as an array is not
equal to "fault" or "maintenance", then include its frequency in an
average of all the PCS's": `IndexFilter(all_pcs_status != "fault" &&
all_pcs_status != "maintenance", all_pcs_frequency)` -- In this example,
`all_pcs_status` and `all_pcs_frequency` would be
`regex(pcs_\\d+_status)` and `regex(pcs_\\d+_frequency)`, respectively.
The assumption is that `all_pcs_status` and `all_pcs_frequency` are the
same length. (If this is not the case, the filter will not evaluate, but
no adverse effects should otherwise be seen.)

Note that this adds two additional functionalities to filter evaluation:
1) Sorting inputs alphanumerically so that template-expanded variable
names are most likely aligned correctly, and 2) Establishing an
evaluation order for filters based on their dependencies. The filter
evaluation order is NOT dependent on filter order in the config, and is
instead evaluated at runtime based on variable dependencies. (This
filter sorting may not be 100% perfect for all applications, but it
seemed to do well in the test cases that I gave it. It is also necessary
if we plan to use loading from DBI in the future, since DBI will not
maintain the order of filters in the original configuration document.)

## How Has This Been Tested?
Created a test script `go_metrics/scripts/unit_test/filters_galore.sh`
and corresponding test config
`go_metrics/examples/unit_test/filters_galore.sh`. Reran tests in
`go_metrics/scripts/math`, `go_metrics/scripts/echo`, and
`go_metrics/scripts/unit_tests`.


Also, random side note: Did some linting while `gopls` is randomly
working for me. (I guess it's in a good mood.)

Possible release notes: Added IndexFilter filtering option.