DC-541 | UI | Alert Management Tab (#625)
## [DC-541 | UI | Alert Management
Tab](https://flexgen.atlassian.net/browse/DC-541)

## Description
Create the initial structure and table to for the Alerts Management tab
of the Activity Log page. On this page, users are able to view all
alerts that have been configured for their site/fleet. The data
displayed on this page represents what comes back on the FIMS URI
`/events/alerts/management`. This PR only covers the display of existing
alert configuration objects. The creation/editing of alert configuration
objects will be handled in a separate ticket.

## How Has This Been Tested?
Manually tested in the UI