DC-537 expose hardware summary stats (#601)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/DC-537)

## Description
This ticket implements retrieval and calculation of some hardware
summary items: the % cpu and mem usage of the system, a timestamp of the
last restart of the system, and a time duration of how long the system
as been running for.

This ticket also fixes an issue in polling for process specific cpu and
mem usage. Originally, the design for polling these values interrupted
the polling mechanism of publishing hardware data to FIMS. I will
explain in further detail with some comments on the review itself about
specifics. This was a necessary code change, and improvement, over how
it was done originally.

## How Has This Been Tested?
Several items to verify:
- make sure the information on `/cops/stats/system/summary` exists, and
is accurate, to the system information. Compare against commands like
`top` for example.
- Make sure that process specific cpu and mem % usage are accurate also.
Verify `/cops/stats/<process>` for a few processes to validate the
adjustment in this reporting is still accurate. Currently, there was a
line of code blocking these values from being updated properly on dev.
This ticket should fix this issue as well.
- Interval between pubs for `/cops/stats/system/summary` is also
correct, and consistently publishes on the defined set interval
(currently set to 3 seconds)