DC 269 Asymmetric ramps (#634)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/DC-269)

## Description

So PGE wants to be able to ramp after a FR event asymmetrically. So when
we enter cool down and begin to ramp down this will provide a second
slew rate that performs the asymmetric ramp.

There is only one slew_object, but depending on which state we are going
into we will ramp differently.
See this image for what this pr does. 

THIS IS AN OPTIONAL configuration. Fully backwards.


## How Has This Been Tested?
`pytest -x -k asymmetric_slew`