ACO-304 GCOM DNP3 support for TLS (#645)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/ACO-304)

## Description
The TMW license that we have purchased does not include DNP3 Security
Version 6. DNP3 Security Version 5 can be purchased as an upgrade to the
library for an additional $6500. For now, we are opting to only include
TLS security, but may come back to discuss secure authentication at a
later date.

Certificates are generated using a 2-level authentication chain with
OpenSSL. Certificates will be stored in a predefined location that will
be hard-coded into the codebase.

`"conn_type": "TLS"` will indicate to the system to use TCP with TLS.

## How Has This Been Tested?
Reran basic python test suite using TLS as the connection type. Reran
C++ unit tests.

**Note:** My biggest concern is the location being hardcoded to a
specific location. I don't know where these certificates should be, so I
used what was in the old DNP3 as the default location.