ACO-326 go metrics test suite update to use behave (#665)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/ACO-326)

## Description
Migrated all "real" tests in the `go_metrics/scripts` directory to the
`behave` BDD test framework. It still needs a lot of work, but it meets
basic testing needs and is definitely easier than having to run those
terrible, nasty, no good bash scripts in bulk.

## How Has This Been Tested?
It _is_ the test? (The results come out as good when the test cases work
and they come out as bad if I mis-copied a test over from the scripts.)

(Also, should be really low risk to merge because it doesn't change any
of the code that's in production...just adds some new testing stuff.)

---------

Co-authored-by: Jack Shade <jnshade@ncsu.edu>