DC-460 | UI | Alert Creation/Modification Page (frontend only) (#646)
## [DC-460 | UI | Alert Creation/Modification Page (frontend
only)](https://flexgen.atlassian.net/browse/DC-460)

## Description
Sets up the basic frontend logic to allows users to create, edit,
duplicate, and delete alert configurations. From the Alert Management
tab of the Activity Log page, users should now be able to view and edit
their existing alerts as well as create new alerts.

NOTE: this PR does not include the logic for parsing rules into a
go_metrics friendly 'expression'. So, while the user should be able to
use the component in the Create Alert panel to set up some rule logic,
this rule logic will not get properly parsed and saved as an expression
on the backend. So, when you click in to edit an existing alert
configuration, the component will not display the existing expression.

## How Has This Been Tested?
Manually tested in the UI