DC 567 grid mode init (#655)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/DC-567)

## Description
An issue was found where site_controller would spam 0 at a battery while
Justin Wald was in China. The cause of the bug was that he had not used
one of the sequences commands that initializes the grid_mode_setpoint
fims_object. This fix adds an `allow_set` variable that will prevent the
spam if those functions aren't used.

## How Has This Been Tested?
1. run HybridOS with the run script. DO NOT START SITE.
2. `fims_send -m set -u /components/ess_psm/on_off_grid_mode 1`
    - Here we set the psm value to be NOT zero. 
4. Observe the `grid_mode` for ess_1
5. Observe fims. `fims_listen -m set -u /components/ess_psm`
- This would catch a set 0 the second that `site_controller` started
prior to this ticket
- Now the command can only be sent AFTER you use a grid_mode based
sequences function.
    - So no spam now.
6. Start site_controller and observe grid_mode for ess 1. Should now be
0.