Dc 501 create parent tab for site diagrams (#641)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/DC-501)
## [Link to Jira Ticket #
2](https://flexgen.atlassian.net/browse/DC-503)

## Description
Create a basic parent container that can be filled with the site
diagrams code. This has some boilerplate in it at present, but in
general this ticket is just supposed to include a container with basic
elements that can be customized and situated later.

In total this accomplishes the following:
- Creates a UI flag which can be enabled or disabled to show/hide the
site diagram UI
- Creates a button in "Dashboard" which shows the site diagram when
clicked (DC-501)
- "Roughly renders" site diagram into the site diagram parent container
(DC-503) 

## How Has This Been Tested?
- As the branch stands, building and running yields no observable
differences.

To QA:
- Change the "enable_site_diagram" flag inside of dashboard.json to true
- Build and run
- On the dashboard page, there will be a 3rd option called "Diagram."
When clicked, you will see a basic diagram.