DC-540 | UI | Alert Rule Creation/Modification Logic (#671)
## [DC-540 | UI | Alert Rule Creation/Modification
Logic](https://flexgen.atlassian.net/browse/DC-540?atlOrigin=eyJpIjoiNWZiOWYzOWQ2MTM0NDNmYWJhM2ZmZWZiODUzNTMzOTEiLCJwIjoiaiJ9)

## Description
Updates the web server to parse the rule logic entered in the UI into a
go_metrics-friendly expression. Also, does the reverse - i.e. adds the
logic to parse the go-metrics expression into an object that can be fed
back into the UI rule logic component. Additionally, adds the ability
for the user to enter in a message that describes each individual
condition in the rule logic. Finally, temporarily removes the fields
dealing with templating in order to have a working alerting MVP.

## How Has This Been Tested?
Manually tested in the UI