Aco 354 ess controller memory leak fix (#705)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/ACO-354)

## Description
The ess controller has known memory leaks for when the program shuts
down while there are schedItems on the reqChan, and there is still
reachable memory from a vector that was never deleted in the main
function. This PR removes both of those memory leaks.

## How Has This Been Tested?
This has been tested locally on my machine with the datamap
configurations that I used to find the memory leaks, and with the
configurations from sites smt, sce, and brp_ffra tx10.

When testing using the site configs, valgrind was run on the
ess_controller and the memory leak of the system was compared to the
memory leakage of the system before these changes were implemented. In
each case, multiple tests were run and with the memory improvements,
valgrind found no memory leaks that were "still reachable" whereas
without the changes, 24 bytes were consistently lost in the "still
reachable" category.

For the SCE site configuration, the number of bytes "definitely lost"
and "indirectly lost" varied depending on how long the program was run
for. There number of bytes "definitely lost" and "indirectly lost" were
too varied to determine if the memory protections decreased the number
of bytes lost.

For the tx10 site configuration, the number of bytes "definitely lost"
and "indirectly lost" was consistently lower when running the
ess_controller with the memory protections compared to running the
ess_controller without.

For the smt site configuration, number of bytes "definitely lost" and
"indirectly lost" was reduced to 0 after implementing the memory
protections; all heap blocks were freed and no leaks were possible.