DC-615 Events support templated alerts w/ global templates & orgs (#694)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/DC-615)

## Description
Further updates to events to support alerting.

Includes:
- Global organizations
All endpoints utilize the name of the organization with the exception of
the POST organization which uses name & id.
- Global templates
Template management is handled inside of the alert management endpoints.
- replyto is optional
- A bunch of refactoring
- Improved validation
- Updates the go_metrics interfacing to latest request.

## How Has This Been Tested?
Unit tests! Should cover every endpoint.

To run the tests, do:
```
npm install
npm run test
```

Beyond this, it has been tested with various combinations of queries
with an emphasis on the frontend interface.