ACO-339 add server delay options for testing (#719)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/ACO-339)

## Description
This PR adds the ability to strip the leading `/` from the
gcom_modbus_server name if one exists.

There is also updated documentation from Phil in
`gcom_modbus_interface/doc/core processing concepts.md`

This ticket also expands the options for the compile flag SERVER_DELAY
to allow for `gcom_modbus_server` to add temporary and permanent server
delays by sending a `_bug` request to the server. It also adds a
"connect delay" to simulate systems that are slow to connect. (The
SERVER_DELAY flag must be added in order to include these changes in the
build, and these changes should not be included in the release build.)

## How Has This Been Tested?
The client and server both build, and have changed minimally from what
is in dev, so I'm just assuming it works.

---------

Co-authored-by: Phil Wilshire <pwilshire@flexgen.com>