ACO-381 bugfix for replies to single-variable gets (#734)
## [Link to Jira ticket](https://flexgen.atlassian.net/browse/ACO-381)

## Description
Replies to single-variable gets was causing a deadlock because
outputScopeMutex was never unlocked. The results of single-variable gets
was also inconsistent with what was being published, so this PR updates
that to use the same message construction function. Also contains a
small bugfix for the process name never being able to be overwritten by
the "name" field in "meta".

### Are there config changes?
- [ ] Yes
  - [ ] Update sandbox configs
  - [ ] Update reference designs
  - [ ] Add updater_tool PR
- [X] No

### How has this been tested?
Manually tested against the offending config (see Jira ticket), and also
re-ran behave tests to make sure that everything still passed.

## After approval, before requesting merge
- [ ] Update Jira ticket description
  - [ ] include config changes as needed
- [ ] Update API documentation _(if this is a backport, ensure you edit
the correct version)_