DC-524 Site Diagrams config (#718)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/DC-524)

## Description
Create `site_diagram.json` to be consumed by the frontend to render site
diagrams.

Adds configurations for both Control and Coordinate (hoth and gargantua
have the same shape as config_dev at present, though should be
differentiated a bit later).

## How Has This Been Tested?
It hasn't been. This will be the initial version, then it can be
modified in the future. Aka the code which will consume this config
hasn't been written yet.

Documentation

https://flexgen.atlassian.net/wiki/spaces/AP/pages/699695105/v12.1.x+WIP+site+diagram.json