DC 418 input source ui toggling (#707)
## [Link to Jira Ticket](https://flexgen.atlassian.net/browse/DC-418)

## Description
Integration has been using metrics and reserved bools to toggle this
stuff via the ui. This wraps the existing input_sources with
fims_objects that are used as controls such that the metrics solution is
obsolete.

## How Has This Been Tested?
There is a pytest that tests that the feature functions as you'd expect.
`pytest -k test_multiple_inputs`

## Should you desire to manually test I would
Take a look at the pytest to see how to modify configs:
1. Take a valid endpoint and give it multiple input control
2. mess with the controls and see which version of the control from 1 is
enabled.
- Make sure the correct value is used based off of the enabled control
from above.