DC 628 slew bug aps (#735)
## [Link to Jira ticket](https://flexgen.atlassian.net/browse/DC-628)

## Description
There was some code that was trying to provide a floor/ceiling behavior
to APS.
This value was getting assigned to `site_kW_demand` and therefore
causeing undesireable behavior.
Reworked the clamping and ensured expected behavior.

I did create a single ESS setup during testing that I don't use in
programmatic testing, but I'm going
to leave this in the pr just because it may be useful later. :shrug:

There is not really any reason to update documentation behavior should
more or less be the same as the original intent.


### How has this been tested?
I have provided testing via pytest.
Please use `pytest -k limits_guantlet`
This will run APS through a variety of use cases and specifically ensure
that standalone poi limits are obeyed.