ACO 371 GCOM dnp3 type error when reading in multi pubs (#727)
## [Link to Jira ticket](https://flexgen.atlassian.net/browse/ACO-371)

## Description
Removed repeated log message "only floats, uints, ints, and bools are
supported..." that was the result of parsing message contents prior to
confirming that the key in the message was actually one of interest.
Code was re-arranged such that we confirm that the point is relevant
prior to checking the value of the key-value pair.

### Are there config changes?
- [ ] Yes
- [ ] Update sandbox configs
- [ ] Update reference designs
- [ ] Add updater_tool PR
- [X] No

### How has this been tested?
Re-ran base python test suite. Re-ran unit tests. Ran the following
commands to confirm functionality in integration_dev:
1. Checkout integration_dev ncemc2/dev/jwald
2. Build site_controller container with docker compose up -d --build
mclauchlin-site-controller-01
3. pkill run.sh from within the container
4. pkill dnp3_server from within the container
5. gcom_dnp3_server
/home/config/dnp3_server/site-controller_fleet-manager_flexgen_dnp3_server.json
6. No error messages should appear after running for a minute or two.

## After approval, before requesting merge
- [X] Update Jira ticket description
- [X] include config changes as needed `n/a`
- [X] Update API documentation _(if this is a backport, ensure you edit
the correct version)_ `n/a`

---------

Co-authored-by: pmiller-flexgen <50499111+pmiller-flexgen@users.noreply.github.com>