adding fims playback utils (#728)
## [Link to Jira ticket] - n/a

## Description
Added some fims playback utilities to the monorepo. These have been
collected from Tony (dtf), Jimmy (fims_replay_python), Graham
(csv_playback), and Stephanie (fims_replay_cpp).

### Are there config changes?
- [ ] Yes
- [ ] Update sandbox configs
- [ ] Update reference designs
- [ ] Add updater_tool PR
- [X] No

### How has this been tested?
Not at all. These are utilities that are not designed for customer
release, so they are dev-only utilities that can be edited as-needed.
Confirmed with Luis Ramos that adding this to
`hybridos/fims/playback_utils` should not affect the build pipeline.

## After approval, before requesting merge
- [X] Update Jira ticket description `n/a`
- [X] include config changes as needed `n/a`
- [X] Update API documentation _(if this is a backport, ensure you edit
the correct version)_ `n/a`