Aco 378 go metrics allow concatenation of string lists for expressions (#755)
## [Link to Jira ticket](https://flexgen.atlassian.net/browse/ACO-378)

## Description
Allow metrics expressions to be lists (instead of/in addition to
strings). This will allow long expressions to be separated into multiple
lines.

### Are there config changes?
- [X] Yes
  - [ ] Update sandbox configs
  - [ ] Update reference designs
  - [ ] Add updater_tool PR
- [ ] No

### How has this been tested?
Updated test suite to pass all `behave` tests (it wasn't doing that with
what was on dev). Updated unit testing to pass all unit tests (it wasn't
doing that with what was on dev. Added a new `behave` unit test to test
new functionality.

## After approval, before requesting merge
- [X] Update Jira ticket description
  - [X] include config changes as needed
- [x] Update API documentation _(if this is a backport, ensure you edit
the correct version)_