Aco 404v2 finalizing brb (#806)
Note that this is the same PR message as
[ACO-404](https://github.com/flexgen-power/hybridos/pull/799)
## [ACO-404](https://flexgen.atlassian.net/browse/ACO-404)
## [ACO-393](https://flexgen.atlassian.net/browse/ACO-393)
## [ACO-412](https://flexgen.atlassian.net/browse/ACO-412)
## [ACO-392](https://flexgen.atlassian.net/browse/ACO-392)
## [ACO-410](https://flexgen.atlassian.net/browse/ACO-410)
## [ACO-399](https://flexgen.atlassian.net/browse/ACO-399)

## Description
This new branch was made because of merge conflicts in the original
ACO-404 PR. This branch was created from dev and then only merged in
changes from hybridos/ess_controller and
hybridos/sandbox/asset_controls. Please confirm that all the necessary
files are where they should be.
I ran a make clean all on this branch successfully, and I ran the
test_uri_IO_BRB script successfully, but have not done any more testing
beyond that.

This PR covers all of the issues linked above. This represents the MVP
for battery rack balancing.

The battery rack balancing algorithm detects racks that have open
contactors, identifies a target votlage to meet, sends power commands to
charge or discharge the closed racks to meet that target voltage
(including a "soft landing" power derate), and automatically flips
contactors and turns the pcs on and off as needed to implement "coarse
balancing"

The system is configured using a config wrapper as described in past
PRs. ACO-404 contains a few updates including adding the option to send
internal remaps as `fimsSet`s so the data flow can be viewed by a
`fimsListen` as well as handling resetting the battery balancing
function if the clear faults button is clicked on the UI while the BRB
function is in an error state.


### Are there config changes?
- [x] Yes
  - [x] Update sandbox configs
  - [ ] Update reference designs
  - [ ] Add updater_tool PR
- [ ] No

### How has this been tested?
This PR has been extensively tested through validation in the lab and
through running locally on docker. See testing scripts in
sandbox/asset_controls/scripts for more detail.

## After approval, before requesting merge
- [x] Update Jira ticket description
  - [x] include config changes as needed
- [ ] Update API documentation _(if this is a backport, ensure you edit
the correct version)_
 


[ACO-404]:
https://flexgen.atlassian.net/browse/ACO-404?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ACO-393]:
https://flexgen.atlassian.net/browse/ACO-393?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ACO-412]:
https://flexgen.atlassian.net/browse/ACO-412?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ACO-392]:
https://flexgen.atlassian.net/browse/ACO-392?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ACO-410]:
https://flexgen.atlassian.net/browse/ACO-410?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[ACO-399]:
https://flexgen.atlassian.net/browse/ACO-399?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ